Add lobby respawn feature
An arena can be configured via /ems set-lobby-respawn <true/false> to allow player re-spawning for players that are within the lobby.
This allows events to be hosted where ems is used to control player re-spawn so rather going back to their normal re-spawn point players are teleported back to the lobby.