Rework how status signs are updated
Rather then accessing the block in a chunk which may or may not be loaded, and if not loaded it may or may not load the chunk and correctly do the update use isChunkLoaded(x,z). This will not load the chunk if the chunk is not already loaded and if it's not loaded then we defer the update till the loading the said chunk (triggered by a player being in the proximity of the chunk).