Implement teleport queue to try and workaround the mass tp issue in Minecraft
If several players all tp in the same tick to an unloaded chunk there is a high probability that at least some players will not see each other due to a bug within the vanilla Minecraft server. To workaround this issue create a queue of players and locations and only teleport players one at a time, leaving a number of ticks in between each teleport.