Make arenas rejoin-able
An arena can be set to allow players to to rejoin it, this also includes handling auto rejoining if the player was in the arena during a server restart or when they logged out.

When a player joins a rejoin-able arena and then leaves it their state (location, selected team, inventory, XP, health etc) will be saved against the arena they are in and their state before they joined the arena will be restored. When they rejoin their state will be restored to what it was when they exited the arena.
Should the server be shut-down while a player is in an arena with rejoin enabled, or should they logout while in such an arena then when they rejoin the server they will automatically rejoin the arena.