Add time limit feature
Allow arenas to have a time limit. Players who have been playing in the arena longer then the time limit will be auto-kicked and will not be allowed to rejoin.

The time limit can be disabled (default), set to a fixed limit or set to a periodic limit. In the case of a periodic limit, every time the period has passed a number the total time limit increases by the specified time limit amount.