Workaround strange behaviour in the Minecraft XP system
There are various methods which can be used to get a players XP each of which seems to work of its own data and these different "views" are not always synchronised which can lead to a players XP being incorrectly set when their state is restored.

This change uses a set of methods which workaround this issue and resolve the known glitch when players can duplicate their XP straight after enchanting.